dom/nodes/Comment-in-doctype.xhtml

<!DOCTYPE html [<!--x-->]><html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>XML: Comment in doctype internal subset</title>
    <link rel="help" href="https://www.w3.org/TR/2004/REC-xml-infoset-20040204/#infoitem.comment"/>
    <!-- "There is a comment information item for each XML comment in the original document,
    except for those appearing in the DTD (which are not represented)." -->
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(() => {
        assert_equals(document.documentElement.previousSibling, document.firstChild);
        assert_equals(document.firstChild.nodeType, Node.DOCUMENT_TYPE_NODE);
      });
    </script>
  </body>
</html>
