LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Comment-in-doctype-expected.txt


FAIL XML: Comment in doctype internal subset assert_equals: expected DocumentType node but got Comment node <!--x-->
